+++ tests/unit/small_text/test_active_learner.py
@@ -126,8 +126,8 @@ class _PoolBasedActiveLearnerTest(object):
         x = np.random.rand(100, 10)
 
         active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, x)
-        self.assertIsNone(active_learner._x_index_to_position)
-        assert_array_equal(np.empty(0, dtype=np.int64), active_learner.x_indices_labeled)
+        self.assertIsNone(active_learner._index_to_position)
+        assert_array_equal(np.empty(0, dtype=np.int64), active_learner.indices_labeled)
         assert_array_equal(np.empty(0, dtype=np.int64), active_learner.y)
 
         y_initial = self._get_y_initial()
@@ -136,19 +136,19 @@ class _PoolBasedActiveLearnerTest(object):
             x_indices_validation = np.random.choice(x_indices_labeled_indices, size=1, replace=False)
 
         active_learner.initialize_data(x_indices_initial, y_initial, retrain=retrain,
-                                       x_indices_ignored=x_indices_ignored,
-                                       x_indices_validation=x_indices_validation)
+                                       indices_ignored=x_indices_ignored,
+                                       indices_validation=x_indices_validation)
 
-        self.assertIsNotNone(active_learner._x_index_to_position)
+        self.assertIsNotNone(active_learner._index_to_position)
         if self.multi_label:
             assert_csr_matrix_equal(y_initial, active_learner.y)
         else:
             assert_array_equal(y_initial, active_learner.y)
-        assert_array_equal(x_indices_initial, active_learner.x_indices_labeled)
+        assert_array_equal(x_indices_initial, active_learner.indices_labeled)
         if x_indices_ignored is not None:
-            assert_array_equal(x_indices_ignored, active_learner.x_indices_ignored)
+            assert_array_equal(x_indices_ignored, active_learner.indices_ignored)
         else:
-            assert_array_equal(np.array([]), active_learner.x_indices_ignored)
+            assert_array_equal(np.array([]), active_learner.indices_ignored)
 
         if retrain:
             clf_factory.new.assert_called_with()
@@ -199,7 +199,7 @@ class _PoolBasedActiveLearnerTest(object):
         active_learner.initialize_data(x_indices_initial, y_initial)
 
         custom_x = np.random.rand(100, 3)
-        active_learner.query(num_samples=num_samples, x=custom_x)
+        active_learner.query(num_samples=num_samples, representation=custom_x)
 
         query_strategy_mock.query.assert_called_with(ANY, custom_x, ANY, x_indices_initial, y_initial,
                                                      n=num_samples)
@@ -220,7 +220,7 @@ class _PoolBasedActiveLearnerTest(object):
         assert custom_x.shape[0] != len(ds)
 
         with self.assertRaises(ValueError):
-            active_learner.query(num_samples=num_samples, x=custom_x)
+            active_learner.query(num_samples=num_samples, representation=custom_x)
 
     def test_query_with_kwargs(self):
 
@@ -245,16 +245,16 @@ class _PoolBasedActiveLearnerTest(object):
     def test_update(self):
         clf_factory = self._get_classifier_factory()
         query_strategy = RandomSampling()
-        x = self._get_random_dataset()
+        dataset = self._get_random_dataset()
 
         with patch.object(PoolBasedActiveLearner, '_retrain', return_value=None) as retrain_mock:
-            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, x)
+            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, dataset)
 
-            x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+            indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
             y_initial = self._get_y_initial()
-            active_learner.initialize_data(x_indices_initial, y_initial)
+            active_learner.initialize_data(indices_initial, y_initial)
 
-            assert_array_equal(x_indices_initial, active_learner.x_indices_labeled)
+            assert_array_equal(indices_initial, active_learner.indices_labeled)
             if self.multi_label:
                 assert_csr_matrix_equal(y_initial, active_learner.y)
             else:
@@ -264,28 +264,28 @@ class _PoolBasedActiveLearnerTest(object):
             y_new = self._get_y_new()
             active_learner.update(y_new)
 
-            assert_array_equal(np.concatenate([x_indices_initial, indices_new]),
-                               active_learner.x_indices_labeled)
+            assert_array_equal(np.concatenate([indices_initial, indices_new]),
+                               active_learner.indices_labeled)
             if self.multi_label:
                 assert_csr_matrix_equal(vstack([y_initial, y_new]), active_learner.y)
             else:
                 assert_array_equal(np.concatenate([y_initial, y_new]), active_learner.y)
 
-        retrain_mock.assert_called_with(x_indices_validation=None)
+        retrain_mock.assert_called_with(indices_validation=None)
 
     def test_update_with_ignored_samples(self):
         clf_factory = self._get_classifier_factory()
         query_strategy = RandomSampling()
-        ds = self._get_random_dataset()
+        dataset = self._get_random_dataset()
 
         with patch.object(PoolBasedActiveLearner, '_retrain', return_value=None) as retrain_mock:
-            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, ds)
+            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, dataset)
 
-            x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+            indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
             y_initial = self._get_y_initial()
-            active_learner.initialize_data(x_indices_initial, y_initial)
+            active_learner.initialize_data(indices_initial, y_initial)
 
-            assert_array_equal(x_indices_initial, active_learner.x_indices_labeled)
+            assert_array_equal(indices_initial, active_learner.indices_labeled)
             if self.multi_label:
                 assert_csr_matrix_equal(y_initial, active_learner.y)
             else:
@@ -304,29 +304,29 @@ class _PoolBasedActiveLearnerTest(object):
             else:
                 y_new = np.array([LABEL_IGNORED, 1, LABEL_IGNORED, 0, LABEL_IGNORED])  # ignore samples 0, 2, and 4
             active_learner.update(y_new)
-            assert_array_equal(np.concatenate([x_indices_initial, [indices_new[1], indices_new[3]]]),
-                               active_learner.x_indices_labeled)
+            assert_array_equal(np.concatenate([indices_initial, [indices_new[1], indices_new[3]]]),
+                               active_learner.indices_labeled)
 
             if self.multi_label:
                 assert_csr_matrix_equal(vstack([y_initial, y_new[[1, 3], :]]), active_learner.y)
             else:
                 assert_array_equal(np.concatenate([y_initial, y_new[[1, 3]]]), active_learner.y)
 
-        retrain_mock.assert_called_with(x_indices_validation=None)
+        retrain_mock.assert_called_with(indices_validation=None)
 
     def test_update_with_only_ignored_samples(self, query_size=5):
         clf_factory = self._get_classifier_factory()
         query_strategy = RandomSampling()
-        ds = self._get_random_dataset()
+        dataset = self._get_random_dataset()
 
         with patch.object(PoolBasedActiveLearner, '_retrain', return_value=None) as retrain_mock:
-            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, ds)
+            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, dataset)
 
-            x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
-            y_initial = ds.y[x_indices_initial]
-            active_learner.initialize_data(x_indices_initial, y_initial)
+            indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+            y_initial = dataset.y[indices_initial]
+            active_learner.initialize_data(indices_initial, y_initial)
 
-            assert_array_equal(x_indices_initial, active_learner.x_indices_labeled)
+            assert_array_equal(indices_initial, active_learner.indices_labeled)
             if self.multi_label:
                 assert_csr_matrix_equal(y_initial, active_learner.y)
             else:
@@ -336,26 +336,27 @@ class _PoolBasedActiveLearnerTest(object):
             y_new = np.array([LABEL_IGNORED]*query_size)  # ignore all samples
             active_learner.update(y_new)
 
-            assert_array_equal(x_indices_initial, active_learner.x_indices_labeled)
+            assert_array_equal(indices_initial, active_learner.indices_labeled)
             assert_labels_equal(y_initial, active_learner.y)
 
-        retrain_mock.assert_has_calls([call(x_indices_validation=None)])
+        retrain_mock.assert_has_calls([call(indices_validation=None)])
 
     def test_update_with_validation_set_given(self):
         clf_factory = self._get_classifier_factory()
         query_strategy = RandomSampling()
-        ds = self._get_random_dataset()
+        dataset = self._get_random_dataset()
 
         with patch.object(PoolBasedActiveLearner, '_retrain', return_value=None) as retrain_mock:
-            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, ds)
+            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, dataset)
 
-            x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+            indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
             y_initial = self._get_y_initial()
-            x_indices_validation = np.random.choice(x_indices_initial, size=10, replace=False)
+            indices_validation = np.random.choice(indices_initial, size=10, replace=False)
 
-            active_learner.initialize_data(x_indices_initial, y_initial, x_indices_validation=x_indices_validation)
+            active_learner.initialize_data(indices_initial, y_initial,
+                                           indices_validation=indices_validation)
 
-            assert_array_equal(x_indices_initial, active_learner.x_indices_labeled)
+            assert_array_equal(indices_initial, active_learner.indices_labeled)
             if self.multi_label:
                 assert_csr_matrix_equal(y_initial, active_learner.y)
             else:
@@ -363,20 +364,20 @@ class _PoolBasedActiveLearnerTest(object):
 
             indices_new = active_learner.query(num_samples=5)
             y_new = self._get_y_new()
-            active_learner.update(y_new, x_indices_validation=x_indices_validation)
-            assert_array_equal(np.concatenate([x_indices_initial, indices_new]),
-                               active_learner.x_indices_labeled)
+            active_learner.update(y_new, indices_validation=indices_validation)
+            assert_array_equal(np.concatenate([indices_initial, indices_new]),
+                               active_learner.indices_labeled)
             if self.multi_label:
                 assert_csr_matrix_equal(vstack([y_initial, y_new]), active_learner.y)
             else:
                 assert_array_equal(np.concatenate([y_initial, y_new]), active_learner.y)
 
-        retrain_mock.assert_has_calls([call(x_indices_validation=x_indices_validation),
-                                       call(x_indices_validation=x_indices_validation)])
+        retrain_mock.assert_has_calls([call(indices_validation=indices_validation),
+                                       call(indices_validation=indices_validation)])
 
     def test_update_reuse_model(self):
         query_strategy = RandomSampling()
-        ds = self._get_random_dataset()
+        dataset = self._get_random_dataset()
 
         clf_mock = ConfidenceEnhancedLinearSVC()
         clf_mock.fit = Mock()
@@ -384,14 +385,14 @@ class _PoolBasedActiveLearnerTest(object):
         clf_factory_mock.new = Mock()
         clf_factory_mock.new.side_effect = [clf_mock, None]
 
-        active_learner = PoolBasedActiveLearner(clf_factory_mock, query_strategy, ds,
+        active_learner = PoolBasedActiveLearner(clf_factory_mock, query_strategy, dataset,
                                                 reuse_model=True)
 
-        x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+        indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
         y_initial = self._get_y_initial()
-        active_learner.initialize_data(x_indices_initial, y_initial)
+        active_learner.initialize_data(indices_initial, y_initial)
 
-        assert_array_equal(x_indices_initial, active_learner.x_indices_labeled)
+        assert_array_equal(indices_initial, active_learner.indices_labeled)
         if self.multi_label:
             assert_csr_matrix_equal(y_initial, active_learner.y)
         else:
@@ -421,23 +422,23 @@ class _PoolBasedActiveLearnerTest(object):
 
     def test_update_duplicate_indices_in_labeled_pool(self):
         query_strategy = RandomSampling()
-        ds = self._get_random_dataset()
+        dataset = self._get_random_dataset()
 
         clf_factory = SklearnClassifierFactory(ConfidenceEnhancedLinearSVC(), self.NUM_CLASSES)
 
-        active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, ds)
+        active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, dataset)
         active_learner._retrain = Mock()
 
-        x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+        indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
         y_initial = self._get_y_initial()
 
-        active_learner.initialize_data(x_indices_initial, y_initial)
+        active_learner.initialize_data(indices_initial, y_initial)
 
         with self.assertRaises(ValueError):
             y_new = np.random.randint(0, 1, size=5)
 
             active_learner.query(num_samples=5)
-            active_learner.queried_indices = x_indices_initial
+            active_learner.indices_queried = indices_initial
             active_learner.update(y_new)
 
         active_learner._retrain.assert_called_once()
@@ -449,106 +450,106 @@ class _PoolBasedActiveLearnerTest(object):
         self._test_update_label_at(retrain=True)
 
     def test_update_label_at_with_validation_set_given(self):
-        self._test_update_label_at(x_indices_validation='random')
+        self._test_update_label_at(indices_validation='random')
 
-    def _test_update_label_at(self, retrain=False, x_indices_validation=None):
+    def _test_update_label_at(self, retrain=False, indices_validation=None):
         clf_factory = self._get_classifier_factory()
         query_strategy = RandomSampling()
-        x = self._get_random_dataset()
+        dataset = self._get_random_dataset()
 
         with patch.object(PoolBasedActiveLearner, '_retrain', return_value=None) as retrain_mock:
-            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, x)
+            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, dataset)
 
-            x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+            indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
             y_initial = np.array([0, 1, 0, 1, 0, 1, 0, 1, 0, 1])
-            active_learner.initialize_data(x_indices_initial, y_initial)
+            active_learner.initialize_data(indices_initial, y_initial)
 
             # query & update
             indices_new = active_learner.query(num_samples=5)
             y_new = np.array([0, 0, 0, 1, 1])
-            if x_indices_validation == 'random':
-                x_indices_validation = np.random.choice(x_indices_initial, size=10, replace=False)
-            active_learner.update(y_new, x_indices_validation=x_indices_validation)
+            if indices_validation == 'random':
+                indices_validation = np.random.choice(indices_initial, size=10, replace=False)
+            active_learner.update(y_new, indices_validation=indices_validation)
 
             active_learner.update_label_at(indices_new[0], 1, retrain=retrain)
             assert_array_equal([1, 0, 0, 1, 1], active_learner.y[-5:])
         if retrain:
-            retrain_mock.assert_has_calls([call(x_indices_validation=x_indices_validation),
-                                           call(x_indices_validation=x_indices_validation)])
+            retrain_mock.assert_has_calls([call(indices_validation=indices_validation),
+                                           call(indices_validation=indices_validation)])
 
     def test_remove_label_at(self):
-        self._test_remove_label_at(x_index=-2)
-        self._test_remove_label_at(x_index=-1)
+        self._test_remove_label_at(index_to_remove=-2)
+        self._test_remove_label_at(index_to_remove=-1)
 
     def test_remove_label_at_with_validation_set_given(self):
-        self._test_remove_label_at(x_index=-1, retrain=True, x_indices_validation='random')
+        self._test_remove_label_at(index_to_remove=-1, retrain=True, indices_validation='random')
 
-    def _test_remove_label_at(self, x_index=-5, retrain=False, x_indices_validation=None):
-        ds = self._get_random_dataset(num_samples=100)
+    def _test_remove_label_at(self, index_to_remove=-5, retrain=False, indices_validation=None):
+        dataset = self._get_random_dataset(num_samples=100)
         clf_factory = SklearnClassifierFactory(ConfidenceEnhancedLinearSVC(), self.NUM_CLASSES)
         query_strategy = RandomSampling()
         with patch.object(PoolBasedActiveLearner, '_retrain', return_value=None) as retrain_mock:
-            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, ds)
+            active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, dataset)
 
-            x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
-            y_initial = ds.y[x_indices_initial]
-            active_learner.initialize_data(x_indices_initial, y_initial)
+            indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+            y_initial = dataset.y[indices_initial]
+            active_learner.initialize_data(indices_initial, y_initial)
 
             # query & update
             indices = active_learner.query(num_samples=5)
-            y_new = ds.y[indices]
+            y_new = dataset.y[indices]
             active_learner.update(y_new)
 
-            if x_indices_validation == 'random':
-                x_indices_validation = np.random.choice(x_indices_initial, size=3, replace=False)
+            if indices_validation == 'random':
+                indices_validation = np.random.choice(indices_initial, size=3, replace=False)
 
-            active_learner.remove_label_at(active_learner.x_indices_labeled[x_index],
-                                           retrain=retrain, x_indices_validation=x_indices_validation)
+            active_learner.remove_label_at(active_learner.indices_labeled[index_to_remove],
+                                           retrain=retrain, x_indices_validation=indices_validation)
 
             if isinstance(y_new, csr_matrix):
-                assert_csr_matrix_equal(ds.y[active_learner.x_indices_labeled[-4:]], active_learner.y[-4:])
-                assert_csr_matrix_equal(ds.y[active_learner.x_indices_labeled[:11]], active_learner.y[:11])
+                assert_csr_matrix_equal(dataset.y[active_learner.indices_labeled[-4:]], active_learner.y[-4:])
+                assert_csr_matrix_equal(dataset.y[active_learner.indices_labeled[:11]], active_learner.y[:11])
             else:
-                assert_array_equal(ds.y[active_learner.x_indices_labeled[-4:]], active_learner.y[-4:])
-                assert_array_equal(ds.y[active_learner.x_indices_labeled[:11]], active_learner.y[:11])
+                assert_array_equal(dataset.y[active_learner.indices_labeled[-4:]], active_learner.y[-4:])
+                assert_array_equal(dataset.y[active_learner.indices_labeled[:11]], active_learner.y[:11])
 
             self.assertEqual(14, active_learner.y.shape[0])
-            self.assertEqual(14, active_learner.x_indices_labeled.shape[0])
+            self.assertEqual(14, active_learner.indices_labeled.shape[0])
 
-        retrain_mock.assert_called_with(x_indices_validation=x_indices_validation)
+        retrain_mock.assert_called_with(indices_validation=indices_validation)
 
     def test_ignore_sample_at(self):
         ds = self._get_random_dataset(num_samples=100)
-        x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+        indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
         y_initial = self._get_y_initial()
-        unlabeled_indices = self._get_unlabeled_indices(x_indices_initial)
-        self._test_ignore_sample_without_label_change(ds, x_indices_initial, y_initial, unlabeled_indices[0])
+        unlabeled_indices = self._get_unlabeled_indices(indices_initial)
+        self._test_ignore_sample_without_label_change(ds, indices_initial, y_initial, unlabeled_indices[0])
 
     def test_ignore_sample_at_with_retrain_true(self):
         ds = self._get_random_dataset(num_samples=100)
-        x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+        indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
         y_initial = self._get_y_initial()
-        unlabeled_indices = self._get_unlabeled_indices(x_indices_initial)
+        unlabeled_indices = self._get_unlabeled_indices(indices_initial)
         self._test_ignore_sample_without_label_change(ds,
-                                                      x_indices_initial,
+                                                      indices_initial,
                                                       y_initial,
                                                       unlabeled_indices[0],
                                                       retrain=True)
 
     def test_ignore_sample_at_with_validation_set_given(self):
         ds = self._get_random_dataset(num_samples=100)
-        x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+        indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
         y_initial = self._get_y_initial()
-        unlabeled_indices = self._get_unlabeled_indices(x_indices_initial)
-        x_indices_validation = np.random.choice(x_indices_initial, size=3, replace=False)
-        self._test_ignore_sample_without_label_change(ds, x_indices_initial, y_initial, unlabeled_indices[0],
-                                                      x_indices_validation=x_indices_validation)
+        unlabeled_indices = self._get_unlabeled_indices(indices_initial)
+        x_indices_validation = np.random.choice(indices_initial, size=3, replace=False)
+        self._test_ignore_sample_without_label_change(ds, indices_initial, y_initial, unlabeled_indices[0],
+                                                      indices_validation=x_indices_validation)
 
     def test_ignore_sample_at_with_label_removed(self):
         ds = self._get_random_dataset(num_samples=100)
-        x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+        indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
         y_initial = self._get_y_initial()
-        self._test_ignore_sample_with_label_change(ds, x_indices_initial, y_initial, x_indices_initial[0])
+        self._test_ignore_sample_with_label_change(ds, indices_initial, y_initial, indices_initial[0])
 
     def test_ignore_sample_at_with_label_removed_and_retrain_true(self):
         ds = self._get_random_dataset(num_samples=100)
@@ -558,62 +559,62 @@ class _PoolBasedActiveLearnerTest(object):
 
     def test_ignore_sample_at_with_label_removed_and_retrain_true_and_validation_set_given(self):
         ds = self._get_random_dataset(num_samples=100)
-        x_indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
-        y_initial = ds.y[x_indices_initial]
-        self._test_ignore_sample_with_label_change(ds, x_indices_initial, y_initial, x_indices_initial[0],
-                                                   retrain=True, x_indices_validation=x_indices_initial[1:3])
+        indices_initial = np.random.choice(np.arange(100), size=10, replace=False)
+        y_initial = ds.y[indices_initial]
+        self._test_ignore_sample_with_label_change(ds, indices_initial, y_initial, indices_initial[0],
+                                                   retrain=True, indices_validation=indices_initial[1:3])
 
-    def _get_unlabeled_indices(self, x_indices_initial, num_samples=100):
+    def _get_unlabeled_indices(self, indices_initial, num_samples=100):
 
         mask = np.ones(num_samples, bool)
-        mask[x_indices_initial] = False
+        mask[indices_initial] = False
         indices = np.arange(num_samples)
 
         return indices[mask]
 
-    def _test_ignore_sample_without_label_change(self, dataset, x_indices_initial, y_initial, x_index_to_ignore,
-                                                 retrain=False, x_indices_validation=None):
+    def _test_ignore_sample_without_label_change(self, dataset, indices_initial, y_initial, index_to_ignore,
+                                                 retrain=False, indices_validation=None):
         clf_factory = self._get_classifier_factory()
         query_strategy = RandomSampling()
         with patch.object(PoolBasedActiveLearner, '_retrain', return_value=None) as retrain_mock:
             active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, dataset)
 
-            active_learner.initialize_data(x_indices_initial, y_initial, x_indices_validation=x_indices_validation)
-            active_learner.ignore_sample_at(x_index_to_ignore, retrain=retrain)
+            active_learner.initialize_data(indices_initial, y_initial, indices_validation=indices_validation)
+            active_learner.ignore_sample_at(index_to_ignore, retrain=retrain)
 
-        retrain_mock.assert_called_with(x_indices_validation=x_indices_validation)
-        assert_array_equal(np.array(x_index_to_ignore), active_learner.x_indices_ignored)
+        retrain_mock.assert_called_with(indices_validation=indices_validation)
+        assert_array_equal(np.array(index_to_ignore), active_learner.indices_ignored)
 
         # whenever no labels change only one _retrain call is expected (regardless of the retrain kwarg)
         self.assertEqual(1, retrain_mock.call_count)
-        retrain_mock.assert_has_calls([call(x_indices_validation=x_indices_validation)])
+        retrain_mock.assert_has_calls([call(indices_validation=indices_validation)])
 
-    def _test_ignore_sample_with_label_change(self, dataset, x_indices_initial, y_initial, x_index_to_ignore,
-                                              retrain=False, x_indices_validation=None):
+    def _test_ignore_sample_with_label_change(self, dataset, indices_initial, y_initial, index_to_ignore,
+                                              retrain=False, indices_validation=None):
         clf_factory = self._get_classifier_factory()
         query_strategy = RandomSampling()
         with patch.object(PoolBasedActiveLearner, '_retrain', return_value=None) as retrain_mock:
             active_learner = PoolBasedActiveLearner(clf_factory, query_strategy, dataset)
-            active_learner.initialize_data(x_indices_initial, y_initial, x_indices_validation=x_indices_validation)
-            self.assertEqual(x_indices_initial.shape[0], active_learner.x_indices_labeled.shape[0])
-            self.assertEqual(x_indices_initial.shape[0], active_learner.y.shape[0])
-            assert_array_equal(np.array([]), active_learner.x_indices_ignored)
+            active_learner.initialize_data(indices_initial, y_initial, indices_validation=indices_validation)
+            self.assertEqual(indices_initial.shape[0], active_learner.indices_labeled.shape[0])
+            self.assertEqual(indices_initial.shape[0], active_learner.y.shape[0])
+            assert_array_equal(np.array([]), active_learner.indices_ignored)
 
-            active_learner.ignore_sample_at(x_index_to_ignore, retrain=retrain,
-                                            x_indices_validation=x_indices_validation)
+            active_learner.ignore_sample_at(index_to_ignore, retrain=retrain,
+                                            indices_validation=indices_validation)
 
-        self.assertEqual(x_indices_initial.shape[0] - 1, active_learner.x_indices_labeled.shape[0])
-        self.assertEqual(x_indices_initial.shape[0] - 1, active_learner.y.shape[0])
+        self.assertEqual(indices_initial.shape[0] - 1, active_learner.indices_labeled.shape[0])
+        self.assertEqual(indices_initial.shape[0] - 1, active_learner.y.shape[0])
 
-        assert_array_equal(np.array(x_index_to_ignore), active_learner.x_indices_ignored)
+        assert_array_equal(np.array(index_to_ignore), active_learner.indices_ignored)
 
         if not retrain:
             self.assertEqual(1, retrain_mock.call_count)
-            retrain_mock.assert_has_calls([call(x_indices_validation=x_indices_validation)])
+            retrain_mock.assert_has_calls([call(indices_validation=indices_validation)])
         else:
             self.assertEqual(2, retrain_mock.call_count)
-            retrain_mock.assert_has_calls([call(x_indices_validation=x_indices_validation),
-                                           call(x_indices_validation=x_indices_validation)])
+            retrain_mock.assert_has_calls([call(indices_validation=indices_validation),
+                                           call(indices_validation=indices_validation)])
 
 
 class PoolBasedActiveLearnerSingleLabelTest(unittest.TestCase, _PoolBasedActiveLearnerTest):